A better name and description for the Sensor.requestedReportingFrequency
The name of the Sensor.requestedReportingFrequency internal slot gave
a wrong impression that it was used only to define a reporting
frequency.

Actually, this slot stores the frequency option and its value is used
for these two things:
- (indirectly) set the requested sampling frequency for the associated
  platform sensor.
- set the reporting frequency for the given Sensor object (in practice
  it sets the upper limit because reporting frequency is always capped to
  sampling frequency).

Now the slot is renamed to `frequency` (so that it reflects the sensor
option name) to avoid this confusion.

Fixes #294