Improve wording in the "Sensor garbage collection" session
"Eligible for garbage collection" can sound ambiguous and lead to
implementers thinking this is a step that happens prior to garbage
collection itself, whereas the original intent of the text added in pull
request #248 ("Be explicit about steps taken when a Sensor instance is
GCed") was that a sensor that _is being GCed_ must be deactivated if
necessary.

Fixes #372.