initialize a sensor object: Explicitly require |options| to be a SensorOptions
This algorithm, invoked by the constructor of interfaces deriving from
`Sensor`, was accepting any dictionary. This meant it was possible, for
example, for any member (including `frequency`) to actually have any type
whatsoever.

Be more strict and require `options` to be a `SensorsOptions` instance, or
an instance of a dictionary inheriting from `SensorOptions`.