Fix broken references to WebDriver definitions.
We were hardcoding some dfn URLs `anchors` section that have changed in the
WebDriver spec, resulting in broken links.

Several of those actually became exported definitions in w3c/webdriver#1639,
so we can simply rely on Bikeshed autolinking working for them. In these
cases, remove the hardcoded dfns and adjust the wording in the spec (e.g.
"WebDriver error" -> "error").

Additionally, use the "WebDriver-2" shortname to refer to the spec, as it
points to the current version which includes the exported definitions that
we need.

Finally, also adapt to w3c/webdriver#1345 which renamed the "Handling
Errors" section to "Errors".

Fixes #443.