Export the "specific conditions" requirement and reword paragraph.
Mostly some Bikeshed cleanup as a follow-up to #434. That pull request
removed the `<dfn>` for "mandatory conditions", which continued to be
referenced by the "specific conditions" item in the "can expose sensor
readings" list.

Remove the reference to it and, at the same time, export the `<dfn>` for
"specific conditions" so that it can actually be used by extension
specifications.