Automation: Remove "virtual sensor type" from virtual sensor metadata
This practically untangles the concepts in the Automation section from the
rest of the specification, which helps the Device Orientation API
specification adopt it as discussed in w3c/deviceorientation#122.

Now a sensor type itself can optionally define a "virtual sensor type",
a string. It is the same string used to refer to a given virtual sensor type
in the WebDriver extension commands.

In practice, this means that:
- The per-top-level traversable virtual sensor mapping has become a mapping
  of virtual sensor types to virtual sensor.
- The per-type virtual sensor metadata mapping has become a mapping of
  virtual sensor types to virtual sensor metadata.
- Sensor types that define a virtual sensor type attempt to retrieve a
  virtual sensor from the virtual sensor mapping using this string as key.

There are no user-visible changes or changes to the WebDriver extension
commands, it is only the concepts and data structures used internally in the
specification that have changed a little.

Derived specifications will need to be adjusted accordingly once this pull
request is merged.