editorial: List all data associated to a platform sensor in one place
Previously, part of the data associated with a platform sensor was listed in
the "Concepts" section and the rest was listed in the "Model" section, which
made it confusing to follow.

Keep the definitions listed in the "Concepts" section there, but also
reference them in "Model" so that the latter contains all data associated
with a platform sensor.

This commit is marked as editorial because it only reorganizes existing
definitions.

Related to #463.